+++ src/game-engine/world/action/hooks/index.ts
@@ -1,17 +1,7 @@
 import { actionHookMap } from '@engine/game-server';
 import { QuestKey } from '@engine/config/quest-config';
-import { TaskExecutor, ActionType, ActionStrength } from '@engine/world/action';
-
-
-export interface HookTask<T = any> {
-    canActivate?: <Q = T>(task: TaskExecutor<Q>) => boolean | Promise<boolean>;
-    activate: <Q = T>(task: TaskExecutor<Q>, index?: number) => void | undefined | boolean | Promise<void | undefined | boolean>;
-    onComplete?: <Q = T>(task: TaskExecutor<Q>) => void | Promise<void>;
-    delay?: number; // # of ticks before execution
-    delayMs?: number; // # of milliseconds before execution
-    interval?: number; // # of ticks between loop intervals (defaults to single run task)
-    intervalMs?: number;  // # of milliseconds between loop intervals (defaults to single run task)
-}
+import { ActionStrength, ActionType } from '@engine/world/action';
+import { HookTask } from '@engine/world/action/hooks/task';
 
 
 /**
@@ -68,3 +58,7 @@ export const getActionHooks = <T extends ActionHook>(actionType: ActionType, fil
 export function sortActionHooks<T = any>(actionHooks: ActionHook<T>[]): ActionHook<T>[] {
     return actionHooks.sort(actionHook => actionHook.questRequirement !== undefined ? -1 : 1);
 }
+
+
+export * from './hook-filters';
+export * from './task';
